extract xrp regexps, cleanup
